Seperated BLEDevice out into NobleDevice and WebBLEDevice usingi nterfaces
