+++ docs/poweredup-node.js.html
@@ -93,9 +93,9 @@ var __importStar = (this &amp;&amp; this.__importStar) || function (mod) {
     return result;
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-const bledevice_1 = require("./bledevice");
 const boostmovehub_1 = require("./boostmovehub");
 const duplotrainbase_1 = require("./duplotrainbase");
+const nobledevice_1 = require("./nobledevice");
 const puphub_1 = require("./puphub");
 const pupremote_1 = require("./pupremote");
 const wedo2smarthub_1 = require("./wedo2smarthub");
@@ -193,7 +193,7 @@ class PoweredUP extends events_1.EventEmitter {
         return Object.keys(this._connectedHubs).map((uuid) => this._connectedHubs[uuid]).filter((hub) => hub.name === name);
     }
     async _discoveryEventHandler(peripheral) {
-        const device = new bledevice_1.BLEDevice(peripheral);
+        const device = new nobledevice_1.NobleDevice(peripheral);
         let hub;
         if (await wedo2smarthub_1.WeDo2SmartHub.IsWeDo2SmartHub(peripheral)) {
             hub = new wedo2smarthub_1.WeDo2SmartHub(device, this.autoSubscribe);
@@ -286,7 +286,7 @@ exports.PoweredUP = PoweredUP;
 <span class="jsdoc-message">
 	Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a>
 	
-		on Thu Feb 7th 2019
+		on Mon Mar 18th 2019
 	
 	using the <a href="https://github.com/docstrap/docstrap">DocStrap template</a>.
 </span>
