+++ docs/puphub.js.html
@@ -308,7 +308,7 @@ exports.PUPHub = PUPHub;
 <span class="jsdoc-message">
 	Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a>
 	
-		on Thu Feb 7th 2019
+		on Mon Mar 18th 2019
 	
 	using the <a href="https://github.com/docstrap/docstrap">DocStrap template</a>.
 </span>
