+++ src/puphub.ts
@@ -1,12 +1,12 @@
 import { Peripheral } from "noble";
 
-import { BLEDevice } from "./bledevice";
 import { LPF2Hub } from "./lpf2hub";
 import { Port } from "./port";
 
 import * as Consts from "./consts";
 
 import Debug = require("debug");
+import { IBLEDevice } from "./interfaces";
 const debug = Debug("puphub");
 
 
@@ -39,7 +39,7 @@ export class PUPHub extends LPF2Hub {
     }
 
 
-    constructor (device: BLEDevice, autoSubscribe: boolean = true) {
+    constructor (device: IBLEDevice, autoSubscribe: boolean = true) {
         super(device, autoSubscribe);
         this.type = Consts.HubType.POWERED_UP_HUB;
         this._ports = {
